move base code to lib_base and adjust in feature user
