+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GithubApiApp"
         tools:targetApi="35">
-        <activity android:name=".feature_users.ui.UsersActivity" />
+        <activity android:name=".feature_users.activity.UsersActivity" />
         <activity-alias
             android:name=".MainActivity"
             android:exported="true"
-            android:targetActivity="com.example.githubapiapp.feature_users.ui.UsersActivity">
+            android:targetActivity="com.example.githubapiapp.feature_users.activity.UsersActivity">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
