+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,4 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
+<resources>
     <!-- Base application theme. -->
-    <style name="Theme.GithubApiApp" parent="Theme.MaterialComponents.Light.NoActionBar">
-        <item name="android:statusBarColor">@color/blue</item>
-    </style>
+    <style name="Theme.GithubApiApp" parent="Theme.MaterialComponents.Light.NoActionBar" />
 </resources>
