+++ libs/lib_base/build.gradle.kts
@@ -4,9 +4,14 @@ plugins {
 
 android {
     namespace = "com.example.githubapiapp.lib_base"
+
+    buildFeatures {
+        viewBinding = true
+    }
 }
 
 dependencies {
     implementation(libs.androidx.appcompat)
+    implementation(libs.androidx.core.ktx)
     implementation(libs.material)
 }
