+++ libs/lib_base/src/main/java/com/example/githubapiapp/lib_base/di/ModuleInjection.kt
@@ -1,0 +1,21 @@
+package com.example.githubapiapp.lib_base.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import javax.inject.Qualifier
+
+@Module
+@InstallIn(SingletonComponent::class)
+class ModuleInjection {
+    @IODispatcher
+    @Provides
+    fun provideIODispatcher(): CoroutineDispatcher = Dispatchers.IO
+}
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class IODispatcher
