+++ libs/lib_base/src/main/java/com/example/githubapiapp/lib_base/exception/DefaultExceptionData.kt
@@ -1,4 +1,4 @@
-package com.example.githubapiapp.feature_users.util.exception
+package com.example.githubapiapp.lib_base.exception
 
 data class DefaultExceptionData(
     val code: Int,
