+++ libs/lib_base/src/main/res/values/colors.xml
