+++ libs/lib_base/src/main/res/values/dimen.xml
